Refactor CallVisualizer VideoCall environment initialization
